refactor out duplicate code in dataset creation
